[python][cleanup] Split utililities in varios files inside utils folder
