+++ bindings/python/crocoddyl/multibody/contacts/multiple-contacts.hpp
@@ -15,6 +15,7 @@
 #include <utility>
 #include <string>
 #include "crocoddyl/multibody/contacts/multiple-contacts.hpp"
+#include "python/crocoddyl/utils/map-converter.hpp"
 
 namespace crocoddyl {
 namespace python {
