+++ bindings/python/crocoddyl/multibody/costs/cost-sum.hpp
@@ -15,6 +15,7 @@
 #include <utility>
 #include <string>
 #include "crocoddyl/multibody/costs/cost-sum.hpp"
+#include "python/crocoddyl/utils/map-converter.hpp"
 
 namespace crocoddyl {
 namespace python {
